# Event/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Event API",
    "version": "1.0.0",
    "description": "This API allows users to manage events and address events.",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/event/api"
    }
  ],
  "paths": {
...
    "/addressEvents": {
      "get": {
        "summary": "Get all address events",
        "operationId": "getAllAddressEvents",
        "description": "This endpoint retrieves a list of all address events stored in the system. Address events represent specific locations tied to a certain event. If no address events are found, a 404 error is returned.",
        "tags": [
          "AddressEvents"
        ],
        "responses": {
          "200": {
            "description": "A list of address events retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddressEvent"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No address events found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a new address event",
        "operationId": "createAddressEvent",
        "description": "This endpoint allows you to create a new address event by providing the address details, such as the event ID, road name, road number, postal code, and the localtown. If any required fields are missing, it returns a 400 error.",
        "tags": [
          "AddressEvents"
        ],
        "requestBody": {
          "description": "Address event to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressEventInput"
              },
              "example": {
                "event_id": "bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb",
                "roadNumber": 485,
                "postCode": "94043",
                "road": "Mountain View",
                "localtown": "f58c9b62-91f0-42ec-96f0-7cc679c0718a"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Address event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressEvent"
                },
                "example": {
                  "id": "b3f5d0a9-22a3-45c0-804b-b1e2b028953b",
                  "event_id": "bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb",
                  "roadNumber": 485,
                  "postCode": "94043",
                  "road": "Mountain View",
                  "localtown": "f58c9b62-91f0-42ec-96f0-7cc679c0718a"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing required fields"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/addressEvents/{id}": {
      "get": {
        "summary": "Get a specific address event by ID",
        "operationId": "getAddressEventById",
        "description": "This endpoint allows you to retrieve a specific address event using its unique ID. If no address event is found for the given ID, it returns a 404 error.",
        "tags": [
          "AddressEvents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the address event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Address event found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressEvent"
                }
              }
            }
          },
          "404": {
            "description": "Address event not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "summary": "Update a specific address event by ID",
        "operationId": "updateAddressEvent",
        "description": "This endpoint allows you to update the details of a specific address event by providing the ID. You can modify the road, road number, postal code, and localtown. If the address event is not found, it returns a 404 error.",
        "tags": [
          "AddressEvents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the address event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Updated address event data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressEventInput"
              },
              "example": {
                "event_id": "bbbbbbbb-bbbb-bbbb-bbbb-bbbbbbbbbbbb",
                "roadNumber": 485,
                "postCode": "94043",
                "road": "Mountain View",
                "localtown": "f58c9b62-91f0-42ec-96f0-7cc679c0718a"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Address event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressEvent"
                }
              }
            }
          },
          "404": {
            "description": "Address event not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific address event by ID",
        "operationId": "deleteAddressEvent",
        "description": "This endpoint allows you to delete a specific address event by ID. If the address event is found and deleted, it returns a 204 status code. If the event is not found, it returns a 404 error.",
        "tags": [
          "AddressEvents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the address event to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Address event deleted successfully"
          },
          "404": {
            "description": "Address event not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/routesEvents/{id}": {
      "get": {
        "summary": "Get a specific routes event by ID",
        "operationId": "getRoutesEventById",
        "description": "This endpoint allows you to retrieve a specific routes event using its unique ID. If the routes event does not exist, a 404 error will be returned.",
        "tags": [
          "RoutesEvents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the routes event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Routes event found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutesEvent"
                }
              }
            }
          },
          "404": {
            "description": "Routes event not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "summary": "Update a specific routes event by ID",
        "operationId": "updateRoutesEvent",
        "description": "This endpoint allows you to update the details of a specific routes event, such as latitude, longitude, order, and the related address event ID. If the routes event is not found, a 404 error will be returned.",
        "tags": [
          "RoutesEvents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the routes event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Updated routes event data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutesEventInput"
              },
              "example": {
                "latitude": 37.7749,
                "longitude": -122.4194,
                "order": 2,
                "addressEvent_id": "dddddddd-dddd-dddd-dddd-dddddddddddd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Routes event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutesEvent"
                }
              }
            }
          },
          "404": {
            "description": "Routes event not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific routes event by ID",
        "operationId": "deleteRoutesEvent",
        "description": "This endpoint allows you to delete a specific routes event by ID. If the routes event is not found, a 404 error will be returned. If the deletion is successful, it returns a 204 status code.",
        "tags": [
          "RoutesEvents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the routes event to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Routes event deleted successfully"
          },
          "404": {
            "description": "Routes event not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Bearer token authentication"
      }
    },
    "schemas": {
...
      "AddressEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID of the address event"
          },
          "event_id": {
            "type": "string",
            "description": "ID of the associated event"
          },
          "localtown": {
            "type": "string",
            "description": "Local town for the event"
          },
          "postCode": {
            "type": "string",
            "description": "Postal code for the event address"
          },
          "roadNumber": {
            "type": "string",
            "description": "Road number where the event is located"
          },
          "road": {
            "type": "string",
            "description": "Road where the event is located"
          }
        }
      },
      "AddressEventInput": {
        "type": "object",
        "required": [
          "event_id",
          "localtown",
          "postCode",
          "roadNumber",
          "road"
        ],
        "properties": {
          "event_id": {
            "type": "string",
            "description": "ID of the associated event"
          },
          "localtown": {
            "type": "string",
            "description": "Address for the event"
          },
          "postCode": {
            "type": "string",
            "description": "Postal code for the event address"
          },
          "roadNumber": {
            "type": "string",
            "description": "City where the event is located"
          },
          "road": {
            "type": "string",
            "description": "State where the event is located"
          }
        }
      },
      "RoutesEvent": {
        "type": "object",
        "properties": {
          "routeID": {
            "type": "string",
            "description": "ID of the route event"
          },
          "latitude": {
            "type": "number",
            "description": "Latitude of the route"
          },
          "longitude": {
            "type": "number",
            "description": "Longitude of the route"
          },
          "order": {
            "type": "integer",
            "description": "Order of the route in the sequence"
          },
          "addressEvent_id": {
            "type": "string",
            "description": "ID of the associated address event"
          }
        }
      },
      "RoutesEventInput": {
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "order",
          "addressEvent_id"
        ],
        "properties": {
          "latitude": {
            "type": "number",
            "description": "Latitude of the route"
          },
          "longitude": {
            "type": "number",
            "description": "Longitude of the route"
          },
          "order": {
            "type": "integer",
            "description": "Order of the route in the sequence"
          },
          "addressEvent_id": {
            "type": "string",
            "description": "ID of the associated address event"
          }
        }
      },
...
    }
  }
}